==> jupyter_notebooks/01_DataCollection.ipynb <==
# Data Collection Notebook

Part of CRISP-DM **Data Understanding**

## Objectives

* Fetch data from Kaggle and save as raw data
* Inspect the data and save it under outputs/datasets/collection.

## Inputs

* Kaggle JSON file - the authentication token.
* Raw downloaded data file

## Outputs

* Generate and save dataset: outputs/dataset/collection/ ....

---

# Change working directory

* Notebooks are stored in a subfolder, therefore when running the notebook in the editor, we need to change the working directory from the current folder to the parent folder.

* We access the current directory with os.getcwd()

import os
current_dir = os.getcwd()
current_dir

We want to make the parent of the current directory the new current directory
* os.path.dirname() gets the parent directory
* os.chir() defines the new current directory

os.chdir(os.path.dirname(current_dir))
print("You set a new current directory")

Confirm the new current directory

current_dir = os.getcwd()
current_dir

# Fetch data from Kaggle

* Install Kaggle package to fetch data

! pip3 install kaggle==1.5.12

* In order to fetch the data you will need to have an authentication token from Kaggle available in your workspace directory, it takes the form of a file named 'kaggle.json'.
* In case you don't have it available, please follow the following steps:
* Log into your existing Kaggle account or create a new account and log in.
* Once you are logged in, on top right of the page, click on your user profile picture, then on “Settings” from the dropdown menu. This will take you to your account settings.
* Scroll down to the section of the page called **API**.
* Click **Expire API** Token to remove any previous tokens.
* To create a new token, click on the **Create New API Token** button. It will generate a fresh authentication token and will download a kaggle.json file to your machine.
* Now make transfer the file to your working director and make sure it is names correctly.
* Once you have the file available run the cell below so that the token is recognized in the session.

import os
os.environ['KAGGLE_CONFIG_DIR'] = os.getcwd()
! chmod 600 kaggle.json

* The data set is located at the following URL: https://www.kaggle.com/datasets/codeinstitute/housing-prices-data
* Define the kaggle path to be the path that follows www.kaggle.com/datasets/ in the urls.
* Set the destination folder.
* Download the data.

KaggleDatasetPath = "codeinstitute/housing-prices-data"
DestinationFolder = "inputs/datasets/raw"
! kaggle datasets download -d {KaggleDatasetPath} -p {DestinationFolder}

* Unzip the downloaded file, delete the zip file and delete the kaggle.json file.

! unzip {DestinationFolder}/*.zip -d {DestinationFolder} \
  && rm {DestinationFolder}/*.zip \
  && rm kaggle.json

---

# Load and Inspect the Kaggle Data

import pandas as pd
df = pd.read_csv(f"inputs/datasets/raw/house-prices/house-price/house_prices_records.csv")
df.head()

* DataFrame Summary

df.info()

* Check for missing values.

df.isnull().sum()

* Check for duplicated entries.

df[df.duplicated(subset=None)]

* Check unique values in the columns with non-numeric data type

for col in df:
    if df[col].dtypes=='object':
        print(col, '-', df[col].unique())
    elif df[col].unique().size < 11:
        print(col, '-', df[col].unique().size)

---

### Preliminary assessment of data

* The data set has 1460 entries/rows and 24 columns.
* The columns are a mix of data types, integers, floats and objects.
* 9 columns are missing data to various degrees.
* 4 columns have non-numerical entries and are thus categorical data.
* 3 further columns have only a small number of unique numerical entries, suggesting that they may also be categorical values.
* Further investigation and Data Cleaning suggested for the next Notebook.

---

# Push files to Repo

* Save the data file in a local folder named 'collection' under 'outputs/datasets/' and push it to the repository.

import os
try:
  os.makedirs(name='outputs/datasets/collection') # create a folder for the data output
except Exception as e:
  print(e)

df.to_csv(f"outputs/datasets/collection/house_prices_records.csv",index=False)


* That completes this Notebook.
* The cell outputs can now be cleared.
* Changes to the workspace can be pushed to the GitHub Repo.
